Add extractChanges (will need for QueryModel version of DetailEditing)
